engine: extract esm/kepler_vertex.js
